Move mobile distribution mutation resolvers into dedicated module
More sensible to have a common module that to split them into modules
corresponding to resources (since they overlap)
